ML/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
mistralai==0.0.8
pydantic==2.0.3
python-dotenv==1.0.0
httpx==0.24.1
pymongo==4.4.1
redis==4.6.0
pika==1.3.2
python-jose==3.3.0
passlib==1.7.4
tenacity==8.2.2
prometheus-client==0.17.1
pytest==7.4.0
pytest-asyncio==0.21.1
python-multipart==0.0.6
sentence-transformers==2.2.2
scikit-learn==1.3.0
torch==2.0.1
faiss-cpu==1.7.4 